refactor(core): :recycle: rename i18n file to get-localization
